ISSUE #1788 - List and report moved out from issues/risks to ticket
